refactor: put utilities under `utils` directory
